backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
psycopg2-binary==2.9.9
python-docx==1.1.0
python-multipart==0.0.9
openai==0.28.1

authx==0.7.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4

